requirements.txt
pinecone
tiktoken
langchain-community
bs4